# docs/openapi/platform-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Observability Platform API",
    "description": "File n�y du?c g?p th? c�ng t? /v3/api-docs c?a ingestion-service v� core-app t?i th?i di?m t?o � c� th? l?ch n?u code sau n�y d?i m� kh�ng regenerate l?i; ngu?n s? th?t lu�n l� 2 endpoint /v3/api-docs s?ng.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/v1/alert-rules/{id}": {
      "put": {
        "tags": [
          "Alert Rules"
        ],
        "summary": "Update an alert rule",
        "description": "Updates an existing alert rule. Yêu cầu role: ADMIN, DEVOPS.",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlertRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Alert rule updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponseAlertRuleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Alert rule not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Alert Rules"
        ],
        "summary": "Delete an alert rule",
        "description": "Deletes an alert rule. Yêu cầu role: ADMIN, DEVOPS.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alert rule deleted successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponseVoid"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Alert rule not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/alert-rules": {
      "get": {
        "tags": [
          "Alert Rules"
        ],
        "summary": "List alert rules",
        "description": "Returns a list of alert rules, optionally filtered by service and environment. Yêu cầu role: ADMIN, DEVOPS.",
        "operationId": "list_3",
        "parameters": [
          {
            "name": "service",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved list of alert rules",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponseListAlertRuleResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Alert Rules"
        ],
        "summary": "Create an alert rule",
        "description": "Creates a new alert rule. Note: conditionValue is required if conditionType is PATTERN_MATCH. Yêu cầu role: ADMIN, DEVOPS.",
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlertRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Alert rule created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponseAlertRuleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (missing ADMIN or DEVOPS role)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/alert-rules/{id}/enabled": {
      "patch": {
        "tags": [
          "Alert Rules"
        ],
        "summary": "Enable or disable an alert rule",
        "description": "Toggles the enabled state of an alert rule. Yêu cầu role: ADMIN, DEVOPS.",
        "operationId": "setEnabled",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAlertRuleEnabledRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Alert rule state updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponseAlertRuleResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Alert rule not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChannelRequest": {
        "required": [
          "channelType"
        ],
        "type": "object",
        "properties": {
          "channelType": {
            "type": "string",
            "enum": [
              "SLACK",
              "TELEGRAM",
              "WEBHOOK",
              "WEBSOCKET"
            ]
          },
          "target": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "UpdateAlertRuleRequest": {
        "required": [
          "conditionType",
          "environment",
          "ruleName",
          "serviceId",
          "severity"
        ],
        "type": "object",
        "properties": {
          "ruleName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "serviceId": {
            "type": "string"
          },
          "environment": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          },
          "conditionType": {
            "type": "string",
            "enum": [
              "ERROR_SPIKE",
              "PATTERN_MATCH"
            ]
          },
          "thresholdValue": {
            "type": "integer",
            "format": "int32"
          },
          "windowSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "conditionValue": {
            "type": "string"
          },
          "severity": {
            "type": "string",
            "enum": [
              "CRITICAL",
              "HIGH",
              "MEDIUM"
            ]
          },
          "isEnabled": {
            "type": "boolean"
          },
          "notificationChannels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChannelRequest"
            }
          }
        }
      },
      "UnifiedResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Status code of the response (e.g., SUCCESS, VALIDATION_ERROR, NOT_FOUND)",
            "example": "SUCCESS"
          },
          "data": {
            "type": "object",
            "description": "The main response data payload"
          },
          "message": {
            "type": "string",
            "description": "Message providing more context, usually on errors"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response (ISO-8601)",
            "example": "2026-09-08T10:15:30Z"
          },
          "path": {
            "type": "string",
            "description": "API path that was called",
            "example": "/api/v1/services"
          },
          "fieldErrors": {
            "type": "array",
            "description": "List of validation errors if status is VALIDATION_ERROR",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string",
                "description": "List of validation errors if status is VALIDATION_ERROR"
              },
              "description": "List of validation errors if status is VALIDATION_ERROR"
            }
          }
        },
        "description": "Unified envelope response format for all API calls"
      },
      "AlertRuleChannelResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "channelType": {
            "type": "string",
            "enum": [
              "SLACK",
              "TELEGRAM",
              "WEBHOOK",
              "WEBSOCKET"
            ]
          },
          "target": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "AlertRuleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "ruleName": {
            "type": "string"
          },
          "serviceId": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "conditionType": {
            "type": "string"
          },
          "thresholdValue": {
            "type": "integer",
            "format": "int32"
          },
          "windowSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "conditionValue": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "notificationChannels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlertRuleChannelResponse"
            }
          }
        },
        "description": "The main response data payload"
      },
      "UnifiedResponseAlertRuleResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Status code of the response (e.g., SUCCESS, VALIDATION_ERROR, NOT_FOUND)",
            "example": "SUCCESS"
          },
          "data": {
            "$ref": "#/components/schemas/AlertRuleResponse"
          },
          "message": {
            "type": "string",
            "description": "Message providing more context, usually on errors"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response (ISO-8601)",
            "example": "2026-09-08T10:15:30Z"
          },
          "path": {
            "type": "string",
            "description": "API path that was called",
            "example": "/api/v1/services"
          },
          "fieldErrors": {
            "type": "array",
            "description": "List of validation errors if status is VALIDATION_ERROR",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string",
                "description": "List of validation errors if status is VALIDATION_ERROR"
              },
              "description": "List of validation errors if status is VALIDATION_ERROR"
            }
          }
        },
        "description": "Unified envelope response format for all API calls"
      },
...
      "CreateAlertRuleRequest": {
        "required": [
          "conditionType",
          "environment",
          "ruleName",
          "serviceId",
          "severity"
        ],
        "type": "object",
        "properties": {
          "ruleName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "description": "Name of the alert rule",
            "example": "High Error Rate"
          },
          "serviceId": {
            "type": "string",
            "description": "The target service to monitor",
            "example": "payment-service"
          },
          "environment": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "description": "Environment filter",
            "example": "production"
          },
          "conditionType": {
            "type": "string",
            "description": "Condition type",
            "example": "ERROR_COUNT_THRESHOLD",
            "enum": [
              "ERROR_SPIKE",
              "PATTERN_MATCH"
            ]
          },
          "thresholdValue": {
            "type": "integer",
            "description": "Threshold value",
            "format": "int32",
            "example": 100
          },
          "windowSeconds": {
            "type": "integer",
            "description": "Time window in seconds to evaluate the condition",
            "format": "int32",
            "example": 300
          },
          "conditionValue": {
            "type": "string",
            "description": "Value to compare against or pattern to match",
            "example": "50"
          },
          "severity": {
            "type": "string",
            "description": "Severity level when triggered",
            "example": "CRITICAL",
            "enum": [
              "CRITICAL",
              "HIGH",
              "MEDIUM"
            ]
          },
          "isEnabled": {
            "type": "boolean",
            "description": "Whether the alert rule is active",
            "example": true
          },
          "notificationChannels": {
            "type": "array",
            "description": "Channels to notify when alert triggers",
            "items": {
              "$ref": "#/components/schemas/ChannelRequest"
            }
          }
        },
        "description": "Request payload for creating a new alert rule"
      },
...
      "SetAlertRuleEnabledRequest": {
        "required": [
          "enabled"
        ],
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        }
      },
...
      "UnifiedResponseListAlertRuleResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Status code of the response (e.g., SUCCESS, VALIDATION_ERROR, NOT_FOUND)",
            "example": "SUCCESS"
          },
          "data": {
            "type": "array",
            "description": "The main response data payload",
            "items": {
              "$ref": "#/components/schemas/AlertRuleResponse"
            }
          },
          "message": {
            "type": "string",
            "description": "Message providing more context, usually on errors"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response (ISO-8601)",
            "example": "2026-09-08T10:15:30Z"
          },
          "path": {
            "type": "string",
            "description": "API path that was called",
            "example": "/api/v1/services"
          },
          "fieldErrors": {
            "type": "array",
            "description": "List of validation errors if status is VALIDATION_ERROR",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string",
                "description": "List of validation errors if status is VALIDATION_ERROR"
              },
              "description": "List of validation errors if status is VALIDATION_ERROR"
            }
          }
        },
        "description": "Unified envelope response format for all API calls"
      },
      "UnifiedResponseVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Status code of the response (e.g., SUCCESS, VALIDATION_ERROR, NOT_FOUND)",
            "example": "SUCCESS"
          },
          "data": {
            "type": "object",
            "description": "The main response data payload"
          },
          "message": {
            "type": "string",
            "description": "Message providing more context, usually on errors"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response (ISO-8601)",
            "example": "2026-09-08T10:15:30Z"
          },
          "path": {
            "type": "string",
            "description": "API path that was called",
            "example": "/api/v1/services"
          },
          "fieldErrors": {
            "type": "array",
            "description": "List of validation errors if status is VALIDATION_ERROR",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string",
                "description": "List of validation errors if status is VALIDATION_ERROR"
              },
              "description": "List of validation errors if status is VALIDATION_ERROR"
            }
          }
        },
        "description": "Unified envelope response format for all API calls"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "X-API-Key": {
        "type": "apiKey",
        "description": "API Key authentication. Required for all endpoints.",
        "name": "X-API-Key",
        "in": "header"
      }
    }
  }
}